WINDUP-1338 adding pipe for converting effortLevel enum numbers to short description
