+++ ui/src/main/webapp/src/app/app.module.ts
@@ -83,6 +83,7 @@ import {PackageChartComponent} from "./shared/package-chart/package-chart.compon
 import {ApplicationDetailsService} from "./reports/application-details/application-details.service";
 import {TechnologyTagComponent} from "./reports/technology-tag/technology-tag.component";
 import {PrettyPathPipe} from "./reports/pretty-path.pipe";
+import {EffortLevelPipe} from "./reports/effort-level.enum";
 import {EventBusService} from "./core/events/event-bus.service";
 import {WindupExecutionService} from "./services/windup-execution.service";
 import {SchedulerService} from "./shared/scheduler.service";
@@ -168,6 +169,7 @@ initializeModelMappingData();
         ApplicationDetailsComponent,
         ApplicationIndexComponent,
         PrettyPathPipe,
+        EffortLevelPipe,
 
         // Report components
         PackageChartComponent,
